+++ src/Vector3.cpp
@@ -3,12 +3,12 @@
 Vector3::Vector3() : x(0), y(0), z(0) {}
 Vector3::Vector3(double x, double y, double z) : x(x), y(y), z(z) {}
 
-long double& Vector3::operator[](int index) {
+double& Vector3::operator[](int index) {
     if(index < 0 || index > 2) throw std::out_of_range("Index must be 0-2");
     return (index == 0 ? x : (index == 1 ? y : z));
 }
 
-const long double& Vector3::operator[](int index) const {
+const double& Vector3::operator[](int index) const {
     if(index < 0 || index > 2) throw std::out_of_range("Index must be 0-2");
     return (index == 0 ? x : (index == 1 ? y : z));
 }
@@ -21,11 +21,11 @@ Vector3 Vector3::operator-(const Vector3& other) const {
     return Vector3(x - other.x, y - other.y, z - other.z);
 }
 
-Vector3 Vector3::operator*(long double scalar) const {
+Vector3 Vector3::operator*(double scalar) const {
     return Vector3(x * scalar, y * scalar, z * scalar);
 }
 
-Vector3 Vector3::operator/(long double scalar) const {
+Vector3 Vector3::operator/(double scalar) const {
     return Vector3(x / scalar, y / scalar, z / scalar);
 }
 
@@ -44,7 +44,7 @@ Vector3& Vector3::operator*=(double scalar) {
     return *this;
 }
 
-long double Vector3::dot(const Vector3& other) const {
+double Vector3::dot(const Vector3& other) const {
     return x * other.x + y * other.y + z * other.z;
 }
 
@@ -60,12 +60,12 @@ Vector3 Vector3::elementwiseMultiply(const Vector3& other) const {
     return Vector3(x * other.x, y * other.y, z * other.z);
 }
 
-long double Vector3::magnitude() const {
+double Vector3::magnitude() const {
     return std::sqrt(x * x + y * y + z * z);
 }
 
 Vector3 Vector3::normalize() const {
-    long double mag = magnitude();
+    double mag = magnitude();
     if(mag == 0) throw std::runtime_error("Cannot normalize zero vector");
     return *this / mag;
 }
